chore: refactor stacked avatars into component
